catch_up_commit

+++ Functions/Image IO/readtiff.m
@@ -63,5 +63,6 @@ for i=n_start:n_end
 end
 TifLink.close();
 w = warning ('on','all');
-im2 = single(im1);
+im2 = double(im1);
+% im2 = im1;
 end
